Refactor namespaces to wallpanelproject
