+++ WallPanelApp/src/main/java/org/wallpanelproject/android/AppCompatPreferenceActivity.java
@@ -1,4 +1,4 @@
-package de.rhuber.homedash;
+package org.wallpanelproject.android;
 
 import android.content.res.Configuration;
 import android.os.Bundle;
