+++ WallPanelApp/src/main/java/org/wallpanelproject/android/WallPanel.java
@@ -1,4 +1,4 @@
-package de.rhuber.homedash;
+package org.wallpanelproject.android;
 
 import android.content.Intent;
 import android.app.Application;
