+++ WallPanelApp/src/main/res/layout/activity_browser.xml
@@ -2,7 +2,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="de.rhuber.homedash.BrowserActivity"
+    tools:context="org.wallpanelproject.android.BrowserActivity"
     android:scrollbars="none">
 
     <WebView
